added regression test for union dependent transforms
