+++ .pre-commit-config.yaml
@@ -19,6 +19,7 @@ repos:
     -   id: trailing-whitespace
     -   id: end-of-file-fixer
     -   id: check-added-large-files
+        exclude: "tests/*"
 
 -   repo: https://github.com/asottile/reorder_python_imports
     rev: v2.6.0
