+++ tests/pipeline/test_survey_responses_version_1.py
@@ -1,25 +1,4 @@
-import pandas as pd
 import pytest
-from mimesis.schema import Schema
-
-from cishouseholds.pipeline.input_file_stages import generate_input_processing_function
-from cishouseholds.pipeline.input_file_stages import survey_responses_v1_parameters
-from dummy_data_generation.schemas import get_voyager_1_data_description
-
-
-@pytest.fixture
-def responses_v1_survey_ETL_output(mimesis_field, pandas_df_to_temporary_csv):
-    """
-    Generate dummy survey responses v1 delta.
-    """
-    schema = Schema(schema=get_voyager_1_data_description(mimesis_field, ["ONS00000000"], ["ONS00000000"]))
-    pandas_df = pd.DataFrame(schema.create(iterations=10))
-    csv_file_path = pandas_df_to_temporary_csv(pandas_df, sep="|")
-    processing_function = generate_input_processing_function(
-        **survey_responses_v1_parameters, include_hadoop_read_write=False
-    )
-    processed_df = processing_function(resource_path=csv_file_path.as_posix())
-    return processed_df
 
 
 @pytest.mark.integration
